src/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SocketIO>=5.3
SQLAlchemy>=1.4
python-dotenv>=0.21
PyJWT>=2.0
bcrypt>=4.0
requests>=2.28
Werkzeug>=2.0
APScheduler>=3.10
python-dateutil>=2.8
eventlet>=0.33
python-socketio>=5.8
pyodbc>=4.0
pymssql>=2.2

